% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{Abalone Data}
\format{
An object of class \code{data.frame} with 4177 rows and 9 columns.
}
\usage{
abalone
}
\description{
The well-known abalone dataset. Included as an example dataset for use with model stacking.
See Warwick Nash, T. S. (1994), "Abalone," UCI Machine Learning Repository. https://doi.org/10.24432/C55C7W.
}
\keyword{datasets}
