% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{aucFun}
\alias{aucFun}
\title{Calculate the error under the receiver operator curve and plot the ROC, given a vector of predicitons and a reference vector.}
\usage{
aucFun(pred, reference)
}
\arguments{
\item{pred}{The name of the prediction vector.}

\item{reference}{The name of the reference vector.}
}
\value{
Area under the receiver operator curve. Also plots the ROC as a side effect.
}
\description{
Calculate the error under the receiver operator curve and plot the ROC, given a vector of predicitons and a reference vector.
}
\examples{
AUC <- aucFun(pred, reference)
}

