% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BootLM}
\alias{BootLM}
\alias{bootLM}
\title{Bootstrap LM}
\usage{
bootLM(x, y, boot_iter = 1000L)
}
\arguments{
\item{x}{A numeric matrix of features}

\item{y}{A numeric vector of the response variable}

\item{boot_iter}{An integer specifying the number of bootstrap iterations.}
}
\description{
C++ implementation of the bootstrapping LM loop used in
\link[ensembleR]{fitAggregationFunction_bootLM}. This function is not
intended to be called directly.
}
