% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{compare}
\alias{compare}
\title{Compare results from caret models trained on multiple instances on Domino.}
\usage{
compare(dataset, par, metric)
}
\arguments{
\item{dataset}{A character string specifying the name of the dataset you are training the models on.}

\item{par}{Character string vector specifying the parameters of interest. Examples of parameters include different values of hyperparameters or different caret methods.}

\item{metric}{A character string specifying the metric you wish to rank caret results by.}
}
\value{
data.frame containing results from caret train objects, ranked in descending order according to the chosen accuracy metric.
}
\description{
Assumes results are downloaded as .Rda files to a single machine. Loads Rda and files and rbinds the results objects into a single data.frame for the sake of comparison.
}
\examples{
models <- c("nnet", "rf", "gbm")
compare(dataset = "BreastCancer", par = models, metric = "ROC")
}

