% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensembler.R
\docType{package}
\name{ensembler}
\alias{ensembler}
\alias{ensembler-package}
\title{ensembler: An R package for building ensemble methods with Domino Datalab (www.dominoup.com).}
\description{
Domino Datalab (www.dominoup.com) offers R users a
way to programmatically trigger runs on cloud-hosted infrastructure with its package, domino.
ensemlber extends domino's functionality by integrating with the well-known caret package to enable
users to train models on seperate AWS instances in the cloud. ensembler also offers a clean interface for downloading and combining
Rdata files from Domino.
}
\section{ensembler functions}{

The ensembler functions launcher, loadResults, compare, modelList, predict.ensemble
}

