% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit models}
\usage{
fit_models(X, y, xgb_params = NULL, rf_params = NULL, svm_params = NULL)
}
\description{
Fit models
}
\section{Functions}{
\itemize{
\item \code{fit_models()}: the specified models
TODO: ADD MORE DOCUMENTATION HERE

}}
