% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_char_encoder.R
\name{fit_char_encoder}
\alias{fit_char_encoder}
\alias{predict.char_encoder}
\title{One Hot Encoder for categorical variables}
\usage{
fit_char_encoder(X)

\method{predict}{char_encoder}(obj, X)
}
\description{
One Hot Encoder for categorical variables
}
\section{Functions}{
\itemize{
\item \code{predict(char_encoder)}: Predict method for char_encoder S3 class

}}
