% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit models}
\usage{
fit_models(
  X,
  Y,
  aggregation_method = NULL,
  calibration_method = NULL,
  validation_pct = 0.2,
  ...
)
}
\description{
` Fit all the models + the aggregation function
}
