% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_models.R
\name{fit_rf}
\alias{fit_rf}
\title{Random Forest}
\usage{
fit_rf(X, y, ...)
}
\arguments{
\item{X}{Covariates for fitting}

\item{y}{target variable}

\item{...}{Additional fitting arguments to be passed to ranger function}
}
\description{
Fit a random forest model using the ranger package
}
