% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregationFunction_bootLM.R
\name{fitAggregationFunction_bootLM}
\alias{fitAggregationFunction_bootLM}
\alias{predict.ModelAggregator_bootLM}
\title{Fit Aggregation function by bootLM fit}
\usage{
fitAggregationFunction_bootLM(Y, y_hat, boot_iter = 1000, ...)

\method{predict}{ModelAggregator_bootLM}(object, y_hat, alpha = 0.05, ...)
}
\arguments{
\item{Y}{The true Y values for fitting}

\item{y_hat}{The predictions output from a model. Note that this is passed
as an \eqn{N \times K} matrix, where \eqn{K} is the number of predictors.}

\item{boot_iter}{The number of bootstrap iterations to perform.}

\item{...}{Additional arguments to pass to the individual aggregation fitting methods. Each has their own
set of parameters for customizing the aggregation process according to the methodology
in the model.}

\item{object}{The fitted ModelAggregator_EM function}

\item{alpha}{The significance level for the prediction interval}
}
\description{
Function Fit a standard linear model to the data for aggregation.
The looping for bootstrap iterations is done through Rcpp, alongside using
base R functions for things like Cholesky decomposition and matrix
multiplication.
}
\section{Functions}{
\itemize{
\item \code{predict(ModelAggregator_bootLM)}: S3 prediction method for the
bootLM aggregation function

}}
\note{
This method is not intended to be used on its own, but rather used through
the \link[ensembleR]{fitAggregationFunction} function.
}
