% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregationFunction_EM.R
\name{fitAggregationFunction_EM}
\alias{fitAggregationFunction_EM}
\alias{predict.ModelAggregator_EM}
\title{Fit Aggregation Function EM}
\usage{
fitAggregationFunction_EM(
  Y,
  y_hat,
  tol = 1e-04,
  max_iter = 1000,
  verbose = FALSE,
  ...
)

\method{predict}{ModelAggregator_EM}(object, y_hat, alpha = 0.05, n_trials = 1000, ...)
}
\arguments{
\item{Y}{The true Y values for fitting}

\item{y_hat}{The predictions output from a model. Note that this is passed
as an \eqn{N \times K} matrix, where \eqn{K} is the number of predictors.}

\item{tol}{Minimum distance between successive likelihoods before the algorithm is
considered to have converged}

\item{max_iter}{Maximum number of iterations the EM algorithm is allowed to run}

\item{verbose}{Logical indicator for whether or not to print progress messages}

\item{...}{Additional arguments; unused and added to align with predict interface.}

\item{object}{The fitted ModelAggregator_EM function}

\item{alpha}{The significance level for the prediction interval}

\item{n_trials}{The number of random draws to run for the mixture}
}
\description{
Fit a Normal mixture model on the regressors using the EM algorithm
}
\section{Functions}{
\itemize{
\item \code{predict(ModelAggregator_EM)}: Predict method for the S3 EM Model Aggregation function.

}}
\note{
This method is not intended to be used on its own, but rather used through
the \link[ensembleR]{fitAggregationFunction} function.
}
