% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregationFunction_lm.R
\name{fitAggregationFunction_lm}
\alias{fitAggregationFunction_lm}
\alias{predict.ModelAggregator_lm}
\title{Fit Aggregation function by simple LM fit}
\usage{
fitAggregationFunction_lm(Y, y_hat, ...)

\method{predict}{ModelAggregator_lm}(object, y_hat, alpha = 0.05, ...)
}
\arguments{
\item{Y}{The true Y values for fitting}

\item{y_hat}{The predictions output from a model. Note that this is passed
as an \eqn{N \times K} matrix, where \eqn{K} is the number of predictors.}

\item{...}{Additional arguments to pass to the individual aggregation fitting methods. Each has their own
set of parameters for customizing the aggregation process according to the methodology
in the model.}
}
\description{
Fit a standard linear model to the data for aggregation.
}
\section{Functions}{
\itemize{
\item \code{predict(ModelAggregator_lm)}: predict method for the LM aggregation function

}}
\note{
This method is not intended to be used on its own, but rather used through
the \link[ensembleR]{fitAggregationFunction} function.
}
