% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregationFunction_quantile.R
\name{fitAggregationFunction_quantile}
\alias{fitAggregationFunction_quantile}
\alias{predict.QuantileAggregation}
\title{Fit Aggregation by Quantile Regression}
\usage{
fitAggregationFunction_quantile(Y, y_hat, ...)

\method{predict}{QuantileAggregation}(object, y_hat, alpha = 0.05, ...)
}
\arguments{
\item{Y}{The true Y values for fitting}

\item{y_hat}{The predictions output from a model. Note that this is passed
as an \eqn{N \times K} matrix, where \eqn{K} is the number of predictors.}

\item{...}{Additional arguments to pass to the individual aggregation fitting methods. Each has their own
set of parameters for customizing the aggregation process according to the methodology
in the model.}

\item{object}{The fitted ModelAggregator_EM function}

\item{alpha}{The significance level for the prediction interval. Note that this parameter is
hardcoded at 0.05 for now. A warning is raised if another alpha value is provided.}
}
\description{
Function Fit a quantile regression model to the data for aggregation.
}
\section{Functions}{
\itemize{
\item \code{predict(QuantileAggregation)}: General S3 method for the quantile prediction method

}}
\note{
This method is not intended to be used on its own, but rather used through
the \link[ensembleR]{fitAggregationFunction} function.
}
