% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_calibration.R
\name{fitCalibrator}
\alias{fitCalibrator}
\title{SD Estimator based on Aggregator}
\usage{
fitCalibrator(
  y_true,
  aggregated_preds,
  y_hat,
  sigma_est,
  n_smoother = 100,
  ...
)
}
\description{
SD Estimator based on Aggregator
}
