% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSTANModel.R
\name{fitSTANModel}
\alias{fitSTANModel}
\title{Fit a Normal mixture model on the data using STAN}
\usage{
fitSTANModel(y_hat, Y, ...)
}
\arguments{
\item{y_hat}{The predictions output from a model. Note that this is passed an NxK matrix}

\item{Y}{The true Y values for fitting}

\item{...}{Additional arguments to pass to the rstan::sampling function}
}
\description{
Fit a STAN model to the predicted outputs in order to aggregate the final predictions.
Currently, this uses a small mixture model that assumes (1) simplex weights on each model and (2) constant variance across all models.
TODO: this NEEDS to do something other than assuming constant variance, but I'm unsure what.
}
\examples{
\dontrun{
    # Make three small test models for mixing
    y_hat1 <- rnorm(1000, 3, 1)
    y_hat2 <- rnorm(1000, 3, 1.5)
    y_hat3 <- rnorm(1000, 3, 0.5)

    y_hat <- cbind(y_hat1, y_hat2, y_hat3)
    w <- c(0.5, 0.2, 0.3)
    y <- (y_hat \%*\% w + rnorm(1000))[, 1]

    stan_fit <- fitSTANModel(y, y_hat)
    all_samples <- as.matrix(stan_fit)
    rstan::traceplot(stan_fit)
    hist(all_samples[, 1])
}
}
