% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{launcher}
\alias{launcher}
\title{Programmatically pass arguments to a script for training caret models.}
\usage{
launcher(script, dataset, par)
}
\arguments{
\item{script}{A character string specifying the name of the script you wish to run.}

\item{dataset}{A character string specifying the name of the dataset you are training the models on.}

\item{par}{Character string vector specifying the parameters of interest. Examples of parameters include different values of hyperparameters or different caret methods.}
}
\value{
Response from Domino indicating if runs are succesfully triggered.
}
\description{
Programmatically pass arguments to a script for training caret models.
}
\examples{
models <- c("nnet", "rf", "gbm")
launcher(script = "variable_model.R", dataset = "BreastCancer", par = models)
}

