% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAggregationFunction.R
\name{ll}
\alias{ll}
\title{Log Likelihood function}
\usage{
ll(beta_vect)
}
\description{
Log likelihood method for both numerical optimization and EM
algorithm approaches to finding the optimal mixture.
}
