% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAggregationFunction.R
\name{ll_partial}
\alias{ll_partial}
\title{Log Likelihood function}
\usage{
ll_partial(Y, y_hat)
}
\description{
Log likelihood method for both numerical optimization and EM
algorithm approaches to finding the optimal mixture.
}
