% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{load_caret_list}
\alias{load_caret_list}
\title{Create a caretList object from caret models trained on multiple Domino instances.}
\usage{
load_caret_list(dataset, models)
}
\arguments{
\item{dataset}{A character string specifying the name of the dataset you are training the models on.}

\item{models}{Character string vector specifying the names of the models you wish to load.}
}
\value{
a list of class "caretList" containing caret train objects.
}
\description{
Assumes caret train objects are downloaded as .Rda files to a single machine. Loads Rda files and returns all caret models in a single list.
Returns an object of class "caretList" for the sake of interoperablity with the package caretEnsemble.
}
\examples{
models <- c("nnet", "rf", "gbm")
modelList(dataset = "BreastCancer", models = models)
}

