% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{loadResults}
\alias{loadResults}
\title{Load results from a caret train object saved in an .Rda file.}
\usage{
loadResults(dataset, par)
}
\arguments{
\item{dataset}{A character string specifying the dataset associated with the desired results object.}

\item{par}{A character string specifying the parameter or model associated with the desired results object.}
}
\value{
The 'results' slot from a caret train object.
}
\description{
Loads an Rda file in the enclosing function environment.
Allows for assignment of the Rda file to a variable, without polluting the global environment.
}
\examples{
results <- loadResults("BreastCancer", "nnet")
}

