% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{modelList}
\alias{modelList}
\title{Create a list of caret train objects trained on multiple Domino instances.}
\usage{
modelList(dataset, models)
}
\arguments{
\item{dataset}{A character string specifying the name of the dataset you are training the models on.}

\item{models}{Character string vector specifying the names of the models you wish to load.}
}
\value{
a list of class "ensemble" containing caret train objects.
}
\description{
Assumes caret train objects are downloaded as .Rda files to a single machine. Loads Rda files and returns all caret models in a single list.
}
\examples{
models <- c("nnet", "rf", "gbm")
modelList(dataset = "BreastCancer", models = models)
}

