% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSTANModel.R
\name{plot.fitSTANModel}
\alias{plot.fitSTANModel}
\title{STAN model traceplot}
\usage{
\method{plot}{fitSTANModel}(x, ...)
}
\description{
Plot the traceplot fo the fitSTANModel output. Accepts the same options.
}
