% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{predict.ensemble}
\alias{predict.ensemble}
\title{Predict method for objects of class "ensemble"}
\usage{
predict.ensemble(model_list, newdata, type, outcome)
}
\arguments{
\item{model_list}{An object of class "ensemble" containing a list of caret train objects}

\item{newdata}{A data.frame containing the features you wish to predict on.}

\item{type}{A character string specifying the response; for now only supports "prob"}

\item{outcome}{An integer indicating the column index to select from each model's prediction matrix.}
}
\value{
a matrix of predictions, with each column vector corresponding to a different model.
}
\description{
Predict method for objects of class "ensemble"
}
\examples{
model_list <- modelList(dataset="BreastCancer", models=models)
pred_mat <- predict(model_list, newdata=testing, type="prob", outcome=2)
}

