% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSTANModel.R
\name{predict.fitSTANModel}
\alias{predict.fitSTANModel}
\title{Predict using the STAN model}
\usage{
\method{predict}{fitSTANModel}(y_hat)
}
\description{
Runs predictions for a given set of y_hat. Note that the
dimensionality must be the same as in the original training data.
}
