% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(y, y_pred)
}
\arguments{
\item{y}{A numeric vector of true y values}

\item{y_pred}{A numeric vector of predicted values.}
}
\description{
Helper RMSE function.
}
