% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{stackerReg}
\alias{stackerReg}
\title{Stack a set of models using a regularized logistic regression}
\usage{
stackerReg(models, train_mat, pred_mat, cost, type)
}
\arguments{
\item{models}{character vector specifying which models to stack.}

\item{train_mat}{A matrix containing the scored training data created from calling predict.ensemble on the training data, along with a model list.}

\item{pred_mat}{A matrix containing the scored test data created from calling predict.ensemble on the testing data, along with a model list.}

\item{cost}{An integer specifying the cost parameter of the regularized logistic regression. See package LiblineaR for details.}

\item{type}{An integer specifying the type of regularization. See package LiblineaR for details.}
}
\value{
a numeric vector of probabilties indicating the probability of the target class.
}
\description{
Stack a set of models using a regularized logistic regression
}
\examples{
pred_mat <- predict(model_list, newdata=testing, type="prob", outcome=2)
train_mat <- predict(model_list, newdata=training, type="prob", outcome=2)
log_ensemble <- stackerReg(models=models, train_mat=train_mat, pred_mat=pred_mat, cost=1, type=6)
}

