% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{topComb}
\alias{topComb}
\title{Compare the performance of ensembles of top performing models.}
\usage{
topComb(ordered_acc, pred_mat, reference)
}
\arguments{
\item{ordered_acc}{data.frame with caret train results ordered by accuracy.}

\item{pred_mat}{A matrix containing the scored test data created from calling predict.ensemble on the testing data, along with a model list.}

\item{reference}{A numeric vector containing ground truth values}
}
\value{
a data.frame showing the preformance of the ensemlbe of top n models.
}
\description{
Cumulatively adds models to an ensemble based on their rank accuracy, i.e. a "top n models" approach. Creates and tests ensembles by taking a simple average of the models in the ensemble.
}
\examples{
ordered_acc <- compare(models, dataset=dataset, metric="ROC")
dt_model <- topComb(ordered_acc, pred_mat, reference)
}

