\name{ensembleR-package}
\alias{ensembleR-package}
\alias{ensembleR}
\docType{package}
\title{
\packageTitle{ensembleR}
}
\description{
\packageDescription{ensembleR}
}
\details{

ADD DETAILS

}
\author{
\packageAuthor{ensembleR}

Maintainer: \packageMaintainer{ensembleR}
}
\references{
    INSERT REFS
}
\keyword{ package }
\seealso{
    INSERT SEE ALSOS
}
\examples{
    # notrun
    rnorm(100)
}
