\name{ensembleR-package}
\alias{ensembleR-package}
\alias{ensembleR}
\docType{package}
\title{
\packageTitle{ensembleR}
}
\description{
\packageDescription{ensembleR}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{ensembleR}
\packageIndices{ensembleR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ensembleR}

Maintainer: \packageMaintainer{ensembleR}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
