% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_char_encoder.R
\name{fit_char_encoder}
\alias{fit_char_encoder}
\alias{predict.char_encoder}
\title{One Hot Encoder for categorical variables}
\usage{
fit_char_encoder(X)

\method{predict}{char_encoder}(object, X, ...)
}
\arguments{
\item{X}{The covariate features. Expects a data.frame format, since the
function is searching for character columns.}

\item{object}{A char_encoder object, consistent with the predict interface.}

\item{...}{Unused ellipsis argument. Added for consistency with `predict`
interface.}
}
\description{
This is a thin wrapper around the \link[caret]{dummyVars} function that
fits a one-hot encoder to a dataset. Here, we assume that any character
variables are actually categorical, and then use the full-rank version
of those variables (drop one level).
}
\section{Functions}{
\itemize{
\item \code{predict(char_encoder)}: Predict method for char_encoder S3 class

}}
