% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregationFunction_EM.R,
%   R/aggregationFunction_bootLM.R, R/aggregationFunction_lm.R,
%   R/fitAggregationFunction.R, R/fitAggregationFunction_quantile.R
\name{fitAggregationFunction_EM}
\alias{fitAggregationFunction_EM}
\alias{predict.ModelAggregator_EM}
\alias{fitAggregationFunction_bootLM}
\alias{predict.ModelAggregator_bootLM}
\alias{fitAggregationFunction_lm}
\alias{predict.ModelAggregator_lm}
\alias{fitAggregationFunction}
\alias{fitAggregationFunction_quantile}
\alias{predict.QuantileAggregation}
\title{Fit Aggregation Function EM}
\usage{
fitAggregationFunction_EM(
  Y,
  y_hat,
  tol = 1e-04,
  max_iter = 1000,
  verbose = FALSE
)

\method{predict}{ModelAggregator_EM}(obj, y_hat, alpha = 0.05, n_trials = 1000, ...)

fitAggregationFunction_bootLM(Y, y_hat, boot_iter = 1000, ...)

\method{predict}{ModelAggregator_bootLM}(obj, y_hat, alpha = 0.05, n_trials = 1000, return_sims = FALSE, ...)

fitAggregationFunction_lm(Y, y_hat, calibrate_sd = FALSE, ...)

\method{predict}{ModelAggregator_lm}(obj, y_hat, alpha = 0.05, ...)

fitAggregationFunction(Y, y_hat, method = "EM", ...)

fitAggregationFunction_quantile(
  Y,
  y_hat,
  quantiles = c(0.025, 0.5, 0.975),
  ...
)

\method{predict}{QuantileAggregation}(obj, y_hat, ...)
}
\arguments{
\item{Y}{The true Y values for fitting}

\item{y_hat}{The predictions output from a model. Note that this is passed as an $NxK$ matrix, where $K$ is the number of predictors.}

\item{...}{Additional arguments to pass to the optim function}
}
\description{
Fit a normal mixture that maximizes the log likelihood of the of the weights of
each method. This uses the base R `optim` function to find the optimal weights
for each method.
}
\section{Functions}{
\itemize{
\item \code{fitAggregationFunction_EM()}: Fit a Normal mixture model on the regressors using the EM algorithm

\item \code{predict(ModelAggregator_EM)}: Predict method for the S3 EM Model Aggregation function.

\item \code{fitAggregationFunction_bootLM()}: Fit a standard linear model to the data for aggregation.

\item \code{predict(ModelAggregator_bootLM)}: S3 prediction method for the bootLM aggregation function

\item \code{fitAggregationFunction_lm()}: Fit a standard linear model to the data for aggregation.

\item \code{predict(ModelAggregator_lm)}: predict method for the LM aggregation function

\item \code{fitAggregationFunction_quantile()}: Fit a quantile regression model to the data for aggregation.

\item \code{predict(QuantileAggregation)}: General S3 method for the quantile prediction method

}}
\examples{
\dontrun{
# Make three small test models for mixing
y_hat1 <- rnorm(1000, 3, 1)
y_hat2 <- rnorm(1000, 3, 1.5)
y_hat3 <- rnorm(1000, 3, 0.5)

y_hat <- cbind(y_hat1, y_hat2, y_hat3)

w <- c(0.5, 0.2, 0.3)
y <- (y_hat \%*\% w + rnorm(1000))[, 1]

model_aggregator <- fitAggregationFunction(y, y_hat)
predict(model_aggregator, y_hat, alpha=0.05)
plot(fitAggregationFunction)
}
}
