% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_calibration.R
\name{fitCalibrator}
\alias{fitCalibrator}
\title{Calibrate SD Estimates}
\usage{
fitCalibrator(y_true, aggregated_preds, y_hat, sigma_est, n_smoother = 25, ...)
}
\description{
Take the outputs of the aggregator and using this as a mean, fit
the predicted standard deviation of the residuals.
TODO: EXPAND
}
