% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_func.R
\name{loadResults}
\alias{loadResults}
\title{Load results from a caret train object saved in an .Rda file; can assign output to a variable.}
\usage{
loadResults(dataset, par)
}
\arguments{
\item{dataset}{A character string specifying the dataset associated with the desired results object.}

\item{par}{A character string specifying the parameter or model associated with the desired results object.}
}
\value{
The 'results' slot from a caret train object.
}
\description{
Load results from a caret train object saved in an .Rda file; can assign output to a variable.
}
\examples{
results <- loadResults("BreastCancer", "nnet")
}

